+++ app/Http/Controllers/Product/ProductCategoryController.php
@@ -3,8 +3,8 @@
 namespace App\Http\Controllers\Product;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Product\ProductCategoryRequest;
 use App\Models\Product\ProductCategory;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
@@ -17,16 +17,8 @@ class ProductCategoryController extends Controller
         return view('dashboard.product.category.index', compact('categories', 'category'));
     }
 
-    public function store(Request $request)
+    public function store(ProductCategoryRequest $request)
     {
-        $request->validate([
-            'category_name' => 'required',
-            'category_url' => 'required',
-            'category_description' => 'required',
-            'category_position' => 'required|integer|min:1',
-            'category_icon' => 'required|image|mimes:png,jpg,jpeg',
-        ]);
-
         $slug = Str::slug($request->category_name);
         $icon = $request->file('category_icon');
         $iconPath = $this->assignIconCategory($icon, $slug);
@@ -48,16 +40,8 @@ class ProductCategoryController extends Controller
         return view('dashboard.product.category.edit', compact('category'));
     }
 
-    public function update(Request $request, ProductCategory $category)
+    public function update(ProductCategoryRequest $request, ProductCategory $category)
     {
-        $request->validate([
-            'category_name' => 'required',
-            'category_url' => 'required',
-            'category_description' => 'required',
-            'category_position' => 'required|integer|min:1',
-            'category_icon' => 'image|mimes:png,jpg,jpeg',
-        ]);
-
         $icon = $request->file('category_icon');
 
         if ( $icon ) {
@@ -81,9 +65,13 @@ class ProductCategoryController extends Controller
 
     public function destroy(ProductCategory $category)
     {
-        Storage::delete($category->icon);
-        $category->delete();
-        return redirect()->back()->with("success", "Product category has been deleted.");
+        if ($category->products()->exists()) {
+            return back()->with("error", "This Action is denied. Product Category has related to model.");
+        } else {
+            Storage::delete($category->icon);
+            $category->delete();
+            return back()->with("success", "Product category has been deleted.");
+        }
     }
 
     public function assignIconCategory($icon, $slug)
